+++ src/Utility/BackupExport.php
@@ -52,6 +52,12 @@ class BackupExport
     protected $config;
 
     /**
+     * Default extension
+     * @var string
+     */
+    protected $defaultExtension = 'sql';
+
+    /**
      * Driver containing all methods to export/import database backups
      *  according to the database engine
      * @since 2.0.0
@@ -102,21 +108,25 @@ class BackupExport
 
     /**
      * Sets the compression
-     * @param bool|string $compression Compression type. Supported values are
-     *  `bzip2`, `gzip` and `false` (if you don't want to use compression)
+     * @param bool|string $compression Compression type as string. Supported
+     *  values are `bzip2` and `gzip`. Use `false` for no compression
      * @return \DatabaseBackup\Utility\BackupExport
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupExport-utility#compression
      * @throws InternalErrorException
      * @uses $compression
-     * @uses $driver
+     * @uses $defaultExtension
      * @uses $extension
      */
     public function compression($compression)
     {
-        $this->extension = array_search($compression, $this->driver->getValidCompressions(), true);
+        if ($compression) {
+            $this->extension = array_search($compression, VALID_COMPRESSIONS);
 
-        if (!$this->extension) {
-            throw new InternalErrorException(__d('database_backup', 'Invalid compression type'));
+            if (!$this->extension) {
+                throw new InternalErrorException(__d('database_backup', 'Invalid compression type'));
+            }
+        } else {
+            $this->extension = $this->defaultExtension;
         }
 
         $this->compression = $compression;
@@ -135,7 +145,6 @@ class BackupExport
      * @throws InternalErrorException
      * @uses compression()
      * @uses $config
-     * @uses $driver
      * @uses $filename
      */
     public function filename($filename)
@@ -164,12 +173,12 @@ class BackupExport
         }
 
         //Checks for extension
-        if (!$this->driver->getExtension($filename)) {
+        if (!$this->getExtension($filename)) {
             throw new InternalErrorException(__d('database_backup', 'Invalid file extension'));
         }
 
         //Sets the compression
-        $this->compression($this->driver->getCompression($filename));
+        $this->compression($this->getCompression($filename));
 
         $this->filename = $filename;
 
@@ -211,7 +220,7 @@ class BackupExport
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupExport-utility#export
      * @uses filename()
      * @uses $BackupManager;
-     * @uses $driver
+     * @uses $defaultExtension
      * @uses $emailRecipient
      * @uses $filename
      * @uses $extension
@@ -221,7 +230,7 @@ class BackupExport
     {
         if (empty($this->filename)) {
             if (empty($this->extension)) {
-                $this->extension = $this->driver->getDefaultExtension();
+                $this->extension = $this->defaultExtension;
             }
 
             $this->filename(sprintf('backup_{$DATABASE}_{$DATETIME}.%s', $this->extension));