+++ tests/TestCase/Utility/BackupExportTest.php
@@ -24,7 +24,7 @@ namespace DatabaseBackup\Test\TestCase\Utility;
 
 use Cake\Core\Configure;
 use Cake\Log\Log;
-use Cake\TestSuite\TestCase;
+use DatabaseBackup\TestSuite\TestCase;
 use DatabaseBackup\Utility\BackupExport;
 use Reflection\ReflectionTrait;
 
@@ -81,12 +81,6 @@ class BackupExportTest extends TestCase
         //Deletes debug log
         //@codingStandardsIgnoreLine
         @unlink(LOGS . 'debug.log');
-
-        //Deletes all backups
-        foreach (glob(Configure::read(DATABASE_BACKUP . '.target') . DS . '*') as $file) {
-            //@codingStandardsIgnoreLine
-            @unlink($file);
-        }
     }
 
     /**
@@ -103,6 +97,7 @@ class BackupExportTest extends TestCase
         $this->assertEquals($config['database'], 'test');
         $this->assertEquals($config['driver'], 'Cake\Database\Driver\Mysql');
 
+        $this->assertEquals('sql', $this->getProperty($this->BackupExport, 'defaultExtension'));
         $this->assertInstanceof(DATABASE_BACKUP . '\Driver\Mysql', $this->getProperty($this->BackupExport, 'driver'));
         $this->assertFalse($this->getProperty($this->BackupExport, 'emailRecipient'));
         $this->assertNull($this->getProperty($this->BackupExport, 'extension'));