+++ tests/bootstrap.php
@@ -102,6 +102,7 @@ Cache::setConfig([
 
 ConnectionManager::setConfig('test', ['url' => 'mysql://travis@localhost/test']);
 ConnectionManager::setConfig('test_sqlite', ['url' => 'sqlite:\\' . TMP . 'example.sq3']);
+ConnectionManager::setConfig('test_postgres', ['url' => 'postgres://postgres@localhost/travis_ci_test']);
 
 Configure::write('DatabaseBackup.connection', 'test');
 Configure::write('DatabaseBackup.target', TMP . 'backups');